Added gui-api core

+++ src/main/java/com/creativemd/creativecore/common/utils/ColorUtils.java
@@ -1,6 +1,7 @@
 package com.creativemd.creativecore.common.utils;
 
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.Vec3i;
 
 public class ColorUtils {
 	
@@ -11,11 +12,11 @@ public class ColorUtils {
 		return ((int)color.xCoord & 255) << 16 | ((int)color.yCoord & 255) << 8 | (int)color.zCoord & 255;
 	}
 	
-	public static Vec3d IntToRGB(int color)
+	public static Vec3i IntToRGB(int color)
 	{
 		float r = (float)(color >> 16 & 255);
         float g = (float)(color >> 8 & 255);
         float b = (float)(color & 255);
-        return new Vec3d(r, g, b);
+        return new Vec3i(r, g, b);
 	}
 }
